requirements.txt
Flask==1.0.2
Flask-Caching==1.5.0
gspread==3.1.0
oauth2client==4.1.3
mysql==0.0.2
pprint==0.1
requests==2.21.0
pandas==0.24.2


